backend/requirements.txt
fastapi
uvicorn
python-jose
pymongo
python-dotenv
passlib[bcrypt]